pyproject.toml
[project]
name = "yst-mcp"
version = "0.1.0"
description = "YST KPI Report Collector MCP Service"
requires-python = ">=3.10"
dependencies = [
    "fastmcp>=0.2.0",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "python-dateutil>=2.8.0",
    "playwright>=1.40.0",
]
